refactor(agents): remove self-hosted plan cards from framework fixes NV-8164
Drop ctx.plan, planProgress wire events, and related framework runtime so
self-hosted agents rely on tool approval cards only; keep managed plan
progress types in the API egress layer.
