+++ apps/api/src/app/agents/conversation-runtime/reply/agent-reply.controller.ts
@@ -8,16 +8,11 @@ import { UserSession } from '../../../shared/framework/user.decorator';
 import { AgentReplyPayloadDto } from '../../shared/dtos/agent-reply-payload.dto';
 import { HandleAgentReplyCommand } from './handle-agent-reply/handle-agent-reply.command';
 import { HandleAgentReply } from './handle-agent-reply/handle-agent-reply.usecase';
-import { HandlePlanProgressCommand } from './handle-plan-progress/handle-plan-progress.command';
-import { HandlePlanProgress } from './handle-plan-progress/handle-plan-progress.usecase';
 
 @Controller('/agents')
 @ApiExcludeController()
 export class AgentReplyController {
-  constructor(
-    private handleAgentReply: HandleAgentReply,
-    private handlePlanProgress: HandlePlanProgress
-  ) {}
+  constructor(private handleAgentReply: HandleAgentReply) {}
 
   @Post('/:agentId/reply')
   @HttpCode(HttpStatus.OK)
@@ -28,20 +23,6 @@ export class AgentReplyController {
     @Param('agentId') agentId: string,
     @Body() body: AgentReplyPayloadDto
   ) {
-    if (body.planProgress) {
-      return this.handlePlanProgress.execute(
-        HandlePlanProgressCommand.create({
-          userId: user._id,
-          environmentId: user.environmentId,
-          organizationId: user.organizationId,
-          conversationId: body.conversationId,
-          agentIdentifier: agentId,
-          integrationIdentifier: body.integrationIdentifier,
-          event: body.planProgress as HandlePlanProgressCommand['event'],
-        })
-      );
-    }
-
     return this.handleAgentReply.execute(
       HandleAgentReplyCommand.create({
         userId: user._id,
