+++ packages/framework/src/index.ts
@@ -28,15 +28,6 @@ export type {
   FileRef,
   MessageContent,
   MetadataSignal,
-  PlanControl,
-  PlanHandle,
-  PlanProgressEvent,
-  PlanProgressPhase,
-  PlanStep,
-  PlanStepOpts,
-  PlanStepUpdate,
-  PlanTaskInput,
-  PlanTaskStatus,
   ReplyContent,
   ReplyHandle,
   SentMessageInfo,
