+++ packages/framework/src/resources/agent/index.ts
@@ -39,15 +39,6 @@ export type {
   FileRef,
   MessageContent,
   MetadataSignal,
-  PlanControl,
-  PlanHandle,
-  PlanProgressEvent,
-  PlanProgressPhase,
-  PlanStep,
-  PlanStepOpts,
-  PlanStepUpdate,
-  PlanTaskInput,
-  PlanTaskStatus,
   ReplyContent,
   ReplyHandle,
   SentMessageInfo,
